refactor: keep task config in coordinator
