+++ dimos/control/tasks/cartesian_ik_task/test_cartesian_ik_task.py
@@ -21,8 +21,8 @@ import numpy as np
 import pinocchio
 import pytest
 
+from dimos.control.coordinator import TaskConfig
 from dimos.control.task import CoordinatorState, JointStateSnapshot
-from dimos.control.task_config import TaskConfig
 from dimos.control.tasks.cartesian_ik_task.cartesian_ik_task import (
     CartesianIKTask,
     CartesianIKTaskConfig,
