+++ dimos/control/tasks/eef_twist_task/eef_twist_task.py
@@ -23,8 +23,8 @@ from typing import TYPE_CHECKING
 import numpy as np
 import pinocchio
 
+from dimos.control.coordinator import TaskConfig
 from dimos.control.task import CoordinatorState, JointCommandOutput, ResourceClaim
-from dimos.control.task_config import TaskConfig
 from dimos.control.tasks.cartesian_ik_task.cartesian_ik_task import (
     CartesianIKTask,
     CartesianIKTaskConfig,
