+++ dimos/robot/unitree/go2/blueprints/basic/unitree_go2_keyboard_teleop.py
@@ -24,8 +24,7 @@ Usage:
 from __future__ import annotations
 
 from dimos.control.components import HardwareComponent, HardwareType, make_twist_base_joints
-from dimos.control.coordinator import ControlCoordinator
-from dimos.control.task_config import TaskConfig
+from dimos.control.coordinator import ControlCoordinator, TaskConfig
 from dimos.core.coordination.blueprints import autoconnect
 from dimos.robot.unitree.keyboard_teleop import KeyboardTeleop
